[#27] feat : 로컬에 캐싱하는 배치 제거 (API로 전환)
